requirements.txt
Flask==2.3.3
mysql-connector-python==8.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
Werkzeug==2.3.6
Jinja2==3.1.3
gunicorn==21.2.0
sentence-transformers==2.5.1
faiss-cpu==1.8.0
